Instagram_Bot/requirements.txt
Flask
openai==0.28
requests
instabot
Pillow
instagrapi
